feat(cache): wire the low-hit-rate diagnosis hint into the deck (#267)
The predicate (`diagnose_cache`) has been committed+witness-tested in
stella-model since the original 3 commits, and stats.rs already surfaces it
in the CLI's table/json/csv (a3066f8) — but nothing showed it live in the
deck, where the motivating incident actually happens: a session sitting at
CACHE 0% with no hint anywhere on screen.

- AgentEntry gains cache_is_opt_in_provider/cache_call_count and
  AgentEntry::cache_diagnosis (stella-tui/src/cache_panel.rs), which
  re-implements diagnose_cache's selection logic locally (MIN_TURNS=3,
  hit_rate<threshold, opt-in-and-zero-writes -> OptInNeverEngaged else
  PrefixInstability) rather than depending on stella-model, which the deck
  is deliberately kept free of. is_opt_in_provider is the one piece of real
  domain knowledge (the provider cache-posture taxonomy) that can't be
  cheaply re-derived, so it's resolved once upstream in the CLI's
  cache_insight_for (which already depends on stella-model) and threaded
  through a new Inbound::CacheInsight field, exactly like savings_usd_delta
  and ttl_secs already are.
- render_deck grows the statline to a third row only when the focused
  agent's diagnosis fires (mirrors the existing composer_h data-dependent
  band-height precedent); render_status_bar fills it with
  CacheCause::hint()'s full sentence, byte-identical to what stella stats
  prints, per stella-protocol::cache's documented CLI/TUI wording contract.
- deck.rs's #[cfg(test)] mod tests (774 lines) split into deck_tests.rs
  (#[path], matching stella-cli's agent.rs/agent_tests.rs split) — deck.rs
  had zero file-size-ratchet slack left before this change (already grown
  to the allowlisted+50 ceiling by this branch's earlier commits), so the
  2 new AgentEntry fields needed room made, not just requested.

Witnessed: cache_panel::tests::diagnosis_fires_only_past_min_turns_and_under_threshold,
diagnosis_names_opt_in_absent_vs_prefix_instability (pure predicate), and
deck_render::tests::full_deck_frame_grows_a_third_statline_row_for_a_diagnosed_agent
(end-to-end: a real render_deck call, not just render_status_bar in
isolation, proving the layout actually grows and a healthy session stays
quiet).
